Implemented hashing logic on options to track option state
Options now has a getHash() function which returns a simple hash for
each of the checkboxes using the four lowest bits.

Manager stores this hash after any Options changes are detected. The
Manager-stored hash is queried by Game and stored locally during
initialization and map-handling is determined based on this hash. It
should be much faster than the checkbox DOM queries previously done. Not
that I need the speed but hey, why not be faster, right?